Feat: User CRUD (#6)
* feat(bcrypt): Bcrypt implementation

- Create PasswordService interface for implementation
- Implement bcrypt implementation with PasswordService interface
- Updating the env configuration

* refactor(config): Make config more readable

Make the injection token more readable by defining the file source for each
token

* feat(domain): Update User entity

Adding capabilities to set createdAt and updatedAt properties

* feat(repositories): Implement User Repository

Implementing user repository on mysql module

* feat(domain): Adding user dto

Creating a dto specifically for managing update user process, so the user data
can be updated partially

* feat(exception): Creating NotFoundException

Creating a NotFoundException class to thrown an error as Not Found

* feat(usecase): Implementing usecases for user

Implementing usecases for user

* feat(exception): Add ConflictException

Implement a ConflictException to handle duplicate data errors during
insert and update operations. This ensures proper error reporting when
uniqueness constraints are violated.

* feat(domain): Add createdAt and updatedAt initialization to Image for timestamping

Updated `Image` and `ImageBuilder` to support setting `createdAt` and
`updatedAt`, enabling automatic timestamping during object creation.

* fix(bcrypt): cannot find the bcrypt instance

Fix the code that cannot find the bcrypt instance

* fix(typeorm): cannot find entities on runtime

Cannot find entities, because when in runtime mode the ts file is
generated as js file

* feat(usecase): create user

Adding some verifications to the process of creating a user

* feat(usecase): get user by username

Creating a use case to find the user data based on the username

* feat(usecase): delete user

Creating a use case to delete the user data based on the user id

* feat(repository): find user by username

Adding a function to the repository for searching the user data based on
the username

* feat(repository): implement the image repository

Implement the image repository using typeorm

* feat(controller): implement the user controller

Implement the user constroller

* refactor: Restructure the folder

Restructure the folder the flow of the application

* feat(doc): Implement Swagger UI and scalar

Implementing the swagger UI and the scalar for managing the
documentation of the api

* feat(dto): Create dto for managing the response for user data

Creating a dto file under the application layer to be used as response
format when returning user data

* feat(controller): Define the swagger comment

Define the swagger comment to the controller and the dto file

* feat(dto): UpdateUserDTO

Create a function to mapping UpdateUserRequestDTO to UpdateUserDTO

* feat(repository): Add new function to UserRepository

Adding functions to find the user data by email and by username but
except for specific id

* feat(repository): Injecting ImageRepository

Register the ImageRepository to mysql module so it can be used in other
place

* feat(usecase): UpdateUserUseCase

Adding code to check the email and username data

* feat(controller): Implementing update user process

Implementin the process of updating user in controller

* docs(dto): Give an example value for id

Give an example value for id in parameter

* feat(doc): Update the api documentation

Updating the documentation for Id parameter

* feat(dto): Create IsNotEmptyString decorator

Create a custom decorator for validating if the data is not an empty
string

* feat(usecase): Flow of updating user

Use the id that got from the function to search user data by id instead
the id from parameter to update the user data

* chore(doc): Log the server port

Just adding some code to log the port that used to run the aplication

---------

Co-authored-by: pasha <robby.pasha@salt.co.id>